🚚 move: Dashboard device 폴더 분리
