+++ app/src/components/templates/Dashboard/Tablet/TabletDashboardRow.tsx
