+++ app/src/components/templates/Dashboard/index.tsx
@@ -1,9 +1,9 @@
 import { Center } from '@components/common';
 import { Desktop, Mobile, Tablet } from '@utils/responsive/Device';
 import type { DashboardProps } from '@utils/types/Dashboard';
-import { DesktopDashboard } from './DesktopDashboard';
-import { MobileDashboard } from './MobileDashboard';
-import { TabletDashboard } from './TabletDashboard';
+import { DesktopDashboard } from './Desktop';
+import { MobileDashboard } from './Mobile';
+import { TabletDashboard } from './Tablet';
 
 export const Dashboard = ({
   desktopRows,
