+++ app/src/components/templates/DashboardSkeleton/MobileDashboardSkeleton.tsx
@@ -1,6 +1,6 @@
-import { DashboardItem } from '@components/templates/Dashboard/DashboardItem';
-import { MobileDashboardLayout } from '@components/templates/Dashboard/MobileDashboard';
-import { MobileDashboardRow } from '@components/templates/Dashboard/MobileDashboardRow';
+import { MobileDashboardLayout } from '@components/templates/Dashboard/Mobile';
+import { MobileDashboardRow } from '@components/templates/Dashboard/Mobile/MobileDashboardRow';
+import { DashboardItem } from '@components/templates/Dashboard/shared/DashboardItem';
 import type { MobileDashboardProps } from '@utils/types/Dashboard';
 import { DashboardSkeletonItem } from './DashboardSkeletonItem';
